+++ DiamondKata/Models/RangeSettings.cs
@@ -1,0 +1,15 @@
+/*
+ * Diamond Kata
+ * Version: 1.1.0
+ * Author: Meghal Parikh
+ * Last Updated: 2026-08-24
+ */
+
+namespace DiamondKata.Models;
+
+public sealed class RangeSettings
+{
+    public int LowerLimit { get; init; } = 2;
+
+    public int UpperLimit { get; init; } = 5;
+}